+++ po/POTFILES
@@ -14,8 +14,8 @@ data/ui/presets_manager_window.blp
 data/ui/share_window.blp
 data/ui/welcome_window.blp
 data/ui/window.blp
-gradience/utils/css.py
-gradience/utils/custom_presets.py
+gradience/utils/css_parser.py
+gradience/utils/preset_downloader.py
 gradience/utils/flatpak_overrides.py
 gradience/utils/preset.py
 gradience/utils/repo.py
